Move ObsNode to common and add tests to KeyedIndexedTreeSuite
