+++ common/src/main/scala/explore/model/ObsNode.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2016-2023 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package explore.observationtree
+package explore.model
 
 import cats.Eq
 import cats.Order.*
@@ -30,10 +30,9 @@ enum ObsNode derives Eq:
     case Obs(value) => value.asLeft
     case Grp(group) => group.asRight
 
-  def isObs: Boolean = this match
+  def isObs: Boolean   = this match
     case Obs(_) => true
     case _      => false
-
   def isGroup: Boolean = !isObs
 
 object ObsNode:
